doctor_ai_assistant/requirements.txt
fastapi
uvicorn
sqlalchemy
asyncpg
alembic
pydantic
python-dotenv
openai
qdrant-client
tiktoken
PyPDF2
python-docx
Pillow
pytesseract
python-multipart
agents